+++ src/components/form/form.tsx
@@ -9,7 +9,7 @@ import {
 } from '@stencil/core';
 import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
-import JSONSchemaForm from 'react-jsonschema-form';
+import JSONSchemaForm from '@rjsf/core';
 import retargetEvents from 'react-shadow-dom-retarget-events';
 import { FormError, ValidationStatus } from './form.types';
 import {
@@ -144,7 +144,7 @@ export class Form {
                     liveValidate: true,
                     showErrorList: false,
                     FieldTemplate: FieldTemplate,
-                    ArrayFieldTemplate: ArrayFieldTemplate,
+                    ArrayFieldTemplate: ArrayFieldTemplate as any,
                     ObjectFieldTemplate: ObjectFieldTemplate,
                     formContext: {
                         schema: this.modifiedSchema,
@@ -152,8 +152,8 @@ export class Form {
                         propsFactory: this.propsFactory,
                     },
                     fields: {
-                        SchemaField: CustomSchemaField,
-                        ArrayField: CustomArrayField,
+                        SchemaField: CustomSchemaField as any,
+                        ArrayField: CustomArrayField as any,
                     },
                 },
                 []