+++ src/components/icon/icon.tsx
@@ -12,14 +12,14 @@ import { IconSize } from './icon.types';
  * The size and color of the icon is set in CSS, however there are a few
  * standard sizes defined that can be used with the `size` property.
  *
- * ## Setup
- * * To use **@limetech/lime-icons8**, the `/assets` folder from
- * **@limetech/lime-icons8** must be made available on the webserver.
+ * ### Setup
+ * * To use **@lundalogik/lime-icons8**, the `/assets` folder from
+ * **@lundalogik/lime-icons8** must be made available on the webserver.
  * * To use a different icon set, the icons must be placed in a folder structure
  * that looks like this: `assets/icons/<name-of-icon>.svg`
  *
  * If `assets` is placed in the root, no other setup is needed. The icons will
- * be fetched with a relative URL from `assets/icons/<name-of-icon>.svg`.
+ * be fetched with a relative URL from `/assets/icons/<name-of-icon>.svg`.
  *
  * If `assets` is placed in a sub-folder somewhere, the easiest way to make the
  * icons available is to use the HTML `base` element:
@@ -35,7 +35,20 @@ import { IconSize } from './icon.types';
  * <limel-config config={{iconPath: '/my/parent/path/'}} />
  * ```
  *
- * ## Icon Finder
+ * ### Lime icons
+ * There are icons included in the **@lundalogik/lime-icons8** package which are
+ * designed by our designers at Lime.
+ * The names of these icons start with `-lime-`, which makes them easy to
+ * find using the Icon Finder tool below.
+ * :::note
+ * Some of the `-lime-` icons have multiple colors and use our own CSS variables
+ * instead of HEX or RGB values to visualize their colors. Thus, you must import
+ * our color palette css files into your project to render the icons properly.
+ * Read more about our [Color System](#/DesignGuidelines/color-system.md/)
+ * and how to do this.
+ * :::
+ *
+ * ### Icon Finder
  *
  * Search for an icon and **click on it to copy its name to clipboard**.
  *